cmsBuild: download full chain of Requires and BuildRequires
If we have:

A --(BuildRequires)--> B --(BuildRequires)--> C

We don't need to download C while building A, but it happens to be
that our generated tmpspec-* will have B and C under BuildRequires.

Thus we must download also C, i.e., a full chain on dependencies
(run-time + build-time) as we don't want to modify tmpsec-* and
change package checksum.

Signed-off-by: David Abdurachmanov <davidlt@cern.ch>